+++ projector_installer/run_config.py
@@ -4,7 +4,6 @@
 
 
 """Run configurations related functions"""
-import os
 import shutil
 from os import listdir, rename
 from os.path import join, isdir, basename
@@ -162,10 +161,7 @@ def get_run_config_names(pattern: Optional[str] = None) -> List[str]:
 
 def delete_config(config_name: str) -> None:
     """Removes specified config."""
-    try:
-        rmtree(get_path_to_config(config_name), ignore_errors=False)
-    except OSError:  # Workaround for NFS issue: https://youtrack.jetbrains.com/issue/PRJ-629
-        os.rmdir(get_path_to_config(config_name))
+    rmtree(get_path_to_config(config_name), ignore_errors=False)
 
 
 def rename_config(from_name: str, to_name: str) -> None: