+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Notable changes to this project are documented in this file.
 # Unreleased
 
 ## Added
-- PRJ-681: New option `--uninstall-ide` for `projector config remove` command 
+- PRJ-681: New option `--uninstall-ide` for `projector config remove` command
+- PRJ-669: Allow run multiple IDE instances
 
 ## Fixed
 - PRJ-629: Error with projector config remove on NFS-mounted directory fixed